OnAirの切り替えをリダイレクトではなくAjaxでやる
