+++ app/controllers/admin/live_stream_media_live_controller.rb
@@ -2,7 +2,8 @@ class Admin::LiveStreamMediaLiveController < ApplicationController
   include SecuredAdmin
 
   def bulk_create
-    @conference.tracks.each do |track|
+    # @conference.tracks.each do |track|
+    [Track.find(20), Track.find(21)].each do |track|
       unless track.live_stream_media_live.present?
         CreateMediaLiveJob.perform_later(@conference, track)
       end
