+++ app/views/admin/tracks/_tracks_nav_cndt2021.html.erb
@@ -27,7 +27,7 @@
   </thead>
 
   <tbody id="talks_table_tbody">
-  <%= render 'admin/tracks/partial/talks_table', conference_day: conference_day %>
+  <%= render 'admin/tracks/partial/talks_table', conference_day: conference_day, talks: @talks, track: @track %>
   </tbody>
 </table>
