+++ app/views/admin/tracks/index.js.erb
@@ -1,5 +1,7 @@
+$("#messages").html('<%= j(render partial: 'admin/tracks/partial/messages', locals: {flash: flash}) %>');
+
 recording_group_target = document.getElementById('talks_table_tbody');
-html = "<%= j(render partial: 'admin/tracks/partial/talks_table', locals: {conference_day: @conference_day}) %>"
+html = "<%= j(render partial: 'admin/tracks/partial/talks_table', locals: {conference_day: @conference_day, talks: @talks, track: @track}) %>"
 recording_group_target.innerHTML = html
 
 debug_media_live_information_target = document.getElementById('debug_media_live_information');
