+++ app/views/admin/tracks/partial/_messages.html.erb
@@ -1,0 +1,8 @@
+<% flash.each do |message_type, message| %>
+  <div class="alert alert-<%= alert_type(message_type) %>" role="alert">
+    <%= message %>
+    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+      <span aria-hidden="true">&times;</span>
+    </button>
+  </div>
+<% end %>
